requirements.txt
absl-py==0.15.0
aiohttp==3.6.2
alembic==1.8.0
anyio==3.6.1
apache-airflow==2.3.2
apache-airflow-providers-ftp==2.1.2
apache-airflow-providers-http==2.1.2
apache-airflow-providers-imap==2.2.3
apache-airflow-providers-sqlite==2.1.3
apispec==3.3.2
argcomplete==2.0.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
Babel==2.10.1
blinker==1.4
boto3==1.24.11
botocore==1.27.11
cachelib==0.7.0
cattrs==1.10.0
certifi==2019.9.11
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==7.1.2
clickclick==20.10.2
cloudpickle==2.1.0
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
connexion==2.13.1
cron-descriptor==1.2.24
croniter==1.3.5
cryptography==37.0.2
cycler==0.11.0
Cython==0.29.28
databricks-cli==0.16.7
dataclasses==0.6
datasets==2.2.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
dnspython==2.2.1
docker==5.0.3
docutils==0.18.1
elasticsearch==7.12.0
email-validator==1.2.1
entrypoints==0.4
et-xmlfile==1.1.0
evidently==0.1.51.dev0
filelock==3.7.1
Flask==1.1.4
Flask-AppBuilder==3.4.5
Flask-Babel==2.0.0
Flask-Caching==1.11.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
flatbuffers==1.12
fonttools==4.33.3
fsspec==2022.5.0
gast==0.3.3
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.20
greenlet==1.1.2
grpcio==1.32.0
gunicorn==20.1.0
h11==0.12.0
h5py==2.10.0
httpcore==0.15.0
httpx==0.23.0
huggingface-hub==0.7.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
inflection==0.5.1
install==1.3.5
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.3
jmespath==1.0.0
joblib==1.1.0
jsonschema==4.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.3
lazy-object-proxy==1.7.1
lockfile==0.12.2
loguru==0.6.0
lxml==4.9.1
Mako==1.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.16.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.2
minio==7.1.9
mlflow==1.26.1
multidict==4.7.6
multiprocess==0.70.12.2
numpy==1.23.0
oauthlib==3.2.0
opencc-python-reimplemented==0.1.6
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.2
pathspec==0.9.0
patsy==0.5.2
pendulum==2.1.2
Pillow==9.1.1
plotly==5.8.0
pluggy==1.0.0
prison==0.2.1
prometheus-client==0.14.1
prometheus-flask-exporter==0.20.2
protobuf==4.21.1
psutil==5.9.1
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.12.0
PyJWT==1.7.1
PyMuPDF==1.19.6
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-daemon==2.3.0
python-dateutil==2.8.2
python-docx==0.8.11
python-nvd3==0.15.0
python-pptx==0.6.21
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.1
pytzdata==2020.1
PyYAML==5.4.1
querystring-parser==1.2.4
regex==2022.6.2
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
rfc3986==1.5.0
rich==12.4.4
rsa==4.8
s3transfer==0.6.0
scikit-learn==1.1.1
scipy==1.8.1
setproctitle==1.2.3
#six==1.16.0 
six==1.15.0
smart-open==6.0.0
smmap==5.0.0
sniffio==1.2.0
SQLAlchemy==1.4.9
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
sqlparse==0.4.2
statsmodels==0.13.2
strsimpy==0.2.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-hub==0.12.0
tensorflow-text==2.4.3
termcolor==1.1.0
text-unidecode==1.3
text2vec==1.1.6
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.11.0
tqdm==4.64.0
transformers==4.19.4
typing_extensions==3.7.4
unicodecsv==0.14.1
urllib3==1.26.9
waitress==2.1.2
websocket-client==1.3.2
websockets==8.1
Werkzeug==1.0.1
win32-setctime==1.1.0
wincertstore==0.2
wrapt==1.12.1
WTForms==2.3.3
xgboost==1.2.0
xlrd==2.0.1
XlsxWriter==3.0.3
XMind==1.2.0
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
